% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-gudhi.R
\name{install_gudhi}
\alias{install_gudhi}
\title{Wrapper for installing the GUDHI Python package}
\usage{
install_gudhi(method = "auto", conda = "auto")
}
\arguments{
\item{method}{Installation method. By default, "auto" automatically finds a
method that will work in the local environment. Change the default to force
a specific installation method. Note that the "virtualenv" method is not
available on Windows.}

\item{conda}{The path to a \code{conda} executable. Use \code{"auto"} to allow
\code{reticulate} to automatically find an appropriate \code{conda} binary. See
\strong{Finding Conda} for more details.}
}
\value{

}
\description{
This function uses the \code{\link[reticulate]{py_install}} function to
install the Python version of the GUDHI library.
}
